requirements.txt
# 핵심 웹 프레임워크
fastapi>=0.112.0,<0.113.0
uvicorn>=0.27.0,<0.28.0
pydantic>=2.7.0,<3.0.0

# 템플릿 엔진 및 정적 파일 처리
jinja2>=3.1.3,<4.0.0
aiofiles>=23.2.1,<24.0.0

# 데이터베이스 관련
sqlalchemy>=2.0.27,<3.0.0
alembic>=1.13.1,<2.0.0

# 인증 및 보안
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.7,<0.1.0

# 데이터 처리 및 분석
pandas>=2.1.4,<3.0.0
numpy>=1.26.3,<2.0.0
scikit-learn>=1.4.0,<2.0.0

# API 요청 및 비동기 지원
httpx>=0.26.0,<0.27.0

# 개발 도구 (선택적)
pytest>=7.4.3,<8.0.0
black>=24.1.1,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=7.0.0,<8.0.0
